tcpci_max77759: don't start toggling when tcpci isn't present
In rare cases, the tcpci will complete tcpci_register_port which
will then lead to the tcpm state machine to start running.
When the tcpm starts toggling in an unattached state, a NULL
pointer dereference occurs because the max77759 chip has not
assigned the result of tcpci_register_port to chip->tcpci.

Add a mutex to guard chip->tcpci in max77759_start_toggling.
During probe, take control of the mutex until the assignment
completes.

Test: before assigning chip->tcpci, set a long delay to allow the
tcpm to initialize and attempt to start toggling. This should force
a NULL pointer handling ramdump. Then, implement the fix, and the
ramdump should no longer occur.
Bug: 289535202
Change-Id: Ica222719dca2ee7ea9eda12b6bb13c04e96e12e7
Signed-off-by: RD Babiera <rdbabiera@google.com>