vh_sched: cpu_freq: Applying rate limite logging
In occurence of error for PMU limit, the error log message
keeps printing. This spam log impacts the system performance.
So, we replace the pr_err function with printk_ratelimited to
print out the log.

Also, we check whether cpu is onlining or not before obtaining PMU
event counter values to avoid execution of unncesssary procedure.

In addition, when the PMU limit is disabled, this patch release
the frequency cap faster.

Bug: 301454348
Change-Id: I76e5af71d0e020e270d24ac3dbdda93bd00a8c23
Signed-off-by: Taeju Park <taeju@google.com>