vh: sched: Fix imbalanced adpf counter
Our vendor hooks are loaded after the system has booted and some tasks
are created. This creates a challenge in handling these pre-existing
tasks. One of the challenges is a task could already be enqueued and if
our hooks are registered; then we can end up with a dequeue operation
performed without an equivalent enqueue operation.

We can't register hooks in a stop_machine() as registering a hook is not
atomic operation and will generate warnings.

Make sure we don't allow adpf counter to be decremented only if it is
positive to handle this imbalance.

Bug: 314298578
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: I66c3ba980155ba4599f2bd09c25891bc4464f802
(cherry picked from commit 32988daaf28577517aa2bf1b6684d99dbabfe081)