build_slider.sh: update the SPL for development
The SPL in the platform build is being updated more frequently than
originally anticipated (previously it was updated quarterly in the
trunk_staging builds). Now the trunk staging SPL follows the monthly
SPL. To accomidate this, this patch updates the development build's SPL
to pick the QPR date after the next QPR, e.g. Nov 2023 builds would pick
March 2024.

Note, as a reminder, this is for development purposes only and the
Android platform build will override this SPL.

Tested: verified the SPL is set to '2024-06-05' with this patch
Bug: 274825778
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: Ifb815a285baf9f633c03e122dfd126e653882ec7