arm64/dts: gs201: Add Alive S2MPU
Alive S2MPU is fully trusted and is not managed by the kernel.
However, we need to make sure the kernel can’t map it to use
it in emulation mode of the S2MPU.

So we define it as:
- “deny-all”: So it never gets power managed or updated
- “off-at-boot”: Never configured at boot.

Which mainly means it remains unconfigured but it's MMIO is not
accessible to the kernel.

Bug: 342511931
Signed-off-by: Mostafa Saleh <smostafa@google.com>
(cherry picked from https://partner-android-review.googlesource.com/q/commit:7b40bac463a2ce4ea983b0c5963a8f65532f37fb)
Merged-In: I844cdd4e2485fbae416c618b0b8a83e30b847065
Change-Id: I844cdd4e2485fbae416c618b0b8a83e30b847065